+++ StripeFinancialConnections/StripeFinancialConnections/Source/Helpers/FinancialConnectionsAppearance.swift
@@ -15,7 +15,7 @@ struct FinancialConnectionsAppearance: Equatable {
     private enum ResolvedBrand: Equatable {
         case stripe
         case link
-        case notlink
+        case onelink
     }
 
     struct Colors: Equatable {
@@ -68,9 +68,9 @@ struct FinancialConnectionsAppearance: Equatable {
         case .link:
             self.logo = .link_logo
             self.logoTintColor = Self.linkLogoTintColor(for: theme)
-        case .notlink:
-            self.logo = .notlink_logo
-            self.logoTintColor = .notlinkLogo
+        case .onelink:
+            self.logo = .onelink_logo
+            self.logoTintColor = Self.linkLogoTintColor(for: theme)
         }
     }
 
@@ -81,8 +81,8 @@ struct FinancialConnectionsAppearance: Equatable {
         switch PresentationManager.shared.configuration.linkBrand {
         case .link:
             return .link
-        case .notlink:
-            return .notlink
+        case .onelink:
+            return .onelink
         case .unparsable, .none:
             break
         }
@@ -90,8 +90,8 @@ struct FinancialConnectionsAppearance: Equatable {
         switch brand {
         case .link:
             return .link
-        case .notlink:
-            return .notlink
+        case .onelink:
+            return .onelink
         case .unparsable, .none:
             switch theme {
             case .linkLight:
@@ -149,10 +149,6 @@ extension FinancialConnectionsAppearance.Colors {
 private extension UIColor {
     static let stripeLogo: UIColor = .dynamic(light: .brand600, dark: .neutral0)
     static let linkLogo: UIColor = .dynamic(light: .linkGreen900, dark: .neutral0)
-    static let notlinkLogo: UIColor = .dynamic(
-        light: UIColor(red: 0.23, green: 0.74, blue: 0.66, alpha: 1.0),
-        dark: .neutral0
-    )
 
     // MARK: Neutral
     static var neutral0: UIColor {